requirements.txt
Django==5.0
django-extensions==3.2.3
django-mathfilters==1.0.0
python-dotenv==1.0.1
gunicorn==21.2.0 
psycopg2-binary==2.9.1
dj-database-url==0.5.0
whitenoise == 5.3.0
uvicorn==0.15.0
psycopg